+++ test/FacadeAct.test.ts
@@ -45,7 +45,7 @@ import {
 import snapshotGasCost from './utils/snapshotGasCost'
 import { useEnv } from '#/utils/env'
 
-const DEFAULT_THRESHOLD = fp('0.05') // 5%
+const DEFAULT_THRESHOLD = fp('0.01') // 1%
 
 const createFixtureLoader = waffle.createFixtureLoader